backend/requirements.txt
fastapi>=0.110,<0.120
uvicorn[standard]>=0.27,<0.40
sqlalchemy>=2.0,<3.0
aiosqlite>=0.19,<1.0
asyncpg>=0.29,<1.0
alembic>=1.13,<2.0
python-dateutil>=2.8,<3.0
pydantic>=2.5,<3.0
pydantic-settings>=2.1,<3.0
bcrypt>=4.0,<5.0
python-jose[cryptography]>=3.3,<4.0
email-validator>=2.0,<3.0
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<0.25
httpx>=0.27,<1.0
cryptography>=42.0,<46.0
respx>=0.21,<1.0
